yara-update/requirements.txt
yara-python==4.2.0
GitPython==3.1.27
